+++ packages/riverpod_lint/lib/src/lints/async_value_nullable_pattern.dart
@@ -34,7 +34,11 @@ class AsyncValueNullablePattern extends RiverpodLintRule {
 
       final parent = node.parent;
       // Parent isn't a "final value" pattern
-      if (parent is! PatternField || parent.effectiveName != 'value') return;
+      if (parent is! PatternField ||
+          (parent.effectiveName != 'value' &&
+              parent.effectiveName != 'value2')) {
+        return;
+      }
 
       final grandParent = parent.parent;
       // GrandParent isn't a "AsyncValue(...)"