+++ packages/riverpod/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Unreleased minor
 
+- In 3.0.0, `AsyncValue.value` will be removed and `valueOrNull` will be
+  renamed to `value`.
+  To match this, both are deprecated, and a temporary `AsyncValue.value2` is added.
 - Added various experimental features:
   - A new syntax for providers
   - Mutations