+++ packages/riverpod_lint/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased minor
+
+- Support `AsyncValue.value2`
+
 ## 2.6.5 - 2025-02-28
 
 Upgrade dependencies